+++ tools/xdhq/examples/PYH/Notes/main.py
@@ -116,7 +116,6 @@ def acEdit(notes, dom, id):
 	dom.setLayout("Edit." + index, readAsset( "Note.html") )
 	dom.setContents({ "Title": note['title'], "Description": note['description'] })
 	dom.disableElements(viewModeElements)
-	dom.dressWidgets("Notes")
 	dom.focus("Title")
 
 def acDelete(notes, dom, id):